requirements.txt
asgiref==3.5.2
attrs==22.1.0
cffi==1.15.1
cryptography==38.0.4
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
